+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Send monitoring tests results to Cypress Dashboard
+
 ## [0.4.3] - 2021-05-17
 
 ### Fixed