+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.4] - 2021-05-28
+
+### Changed
+
+- Flaky tests that fail on Saturday to be skipped
+
 ## [0.4.3] - 2021-05-17
 
 ### Fixed